A Pac Man game with a rndomly generated maze using depth first search
This game has three modes: Classic, Random and Crazy.
Classic as name suggests tries to reproduce the classic Pac Man but this time with a health bar.
Random and crazy modes produce a larger maze as levels increase also, they increase the number of ghosts to challenge the player even more.